Move Database to separate file
